[refact] OrderMain 리팩토링
