+++ src/components/Order/OrderHeader/OrderHeader.scss
@@ -5,7 +5,6 @@
   grid-template-columns: 3fr 1fr;
   grid-template-rows: 1fr 3fr;
   grid-column: span 4;
-  grid-row: span 1;
 
   .orderStep {
     @include flexCenter;
