+++ src/pages/Order/Order.scss
@@ -8,172 +8,9 @@
 
   .orderContainer {
     display: grid;
-    grid-template-columns: repeat(4, 1fr);
     grid-template-rows: 1fr 3fr 1fr;
     width: 1000px;
     height: 800px;
-
-    .concertHall {
-      display: grid;
-      grid-template-rows: 1fr 8fr;
-      grid-column: span 3;
-      background-color: $Gray33;
-
-      .stage {
-        @include flexCenter;
-
-        .text {
-          @include flexCenter;
-          width: 90%;
-          height: 80%;
-          margin-left: 2%;
-          border-bottom-left-radius: 50px;
-          border-bottom-right-radius: 50px;
-          font-weight: 700;
-          font-size: 1.4rem;
-          background-color: beige;
-        }
-      }
-
-      .seatTable {
-        display: grid;
-        grid-template-columns: 1fr 30fr;
-        grid-template-rows: 1fr 20fr;
-        gap: 2px;
-        padding: 2%;
-
-        .seatColumn {
-          display: grid;
-          grid-template-columns: repeat(30, 1fr);
-          color: white;
-
-          .column {
-            @include flexCenter;
-            font-size: 0.7rem;
-          }
-        }
-
-        .seatRow {
-          display: grid;
-          grid-template-rows: repeat(20, 1fr);
-          color: white;
-
-          .row {
-            @include flexCenter;
-            font-size: 0.7rem;
-          }
-        }
-
-        .seatList {
-          display: grid;
-          &.column30 {
-            grid-template-columns: repeat(30, 1fr);
-          }
-          &.row20 {
-            grid-template-rows: repeat(20, 1fr);
-          }
-          gap: 1px;
-
-          .seat {
-            width: 19px;
-            height: 19px;
-            border: 1px solid $Gray33;
-
-            &.selected {
-              background-color: $Red;
-              &:hover {
-                background-color: rgba(229, 8, 21, 0.7);
-                cursor: pointer;
-              }
-            }
-
-            &.reserved {
-              background-color: $Black;
-            }
-
-            &.gradeSR {
-              background-color: rgb(242, 159, 6);
-              &:hover {
-                background-color: rgb(242, 159, 6, 0.7);
-                cursor: pointer;
-              }
-            }
-
-            &.gradeR {
-              background-color: rgba(102, 51, 153);
-              &:hover {
-                background-color: rgba(102, 51, 153, 0.7);
-                cursor: pointer;
-              }
-            }
-
-            &.gradeS {
-              background-color: rgba(70, 130, 180);
-              &:hover {
-                background-color: rgba(70, 130, 180, 0.7);
-                cursor: pointer;
-              }
-            }
-
-            &.gradeA {
-              background-color: rgba(153, 153, 153);
-              &:hover {
-                background-color: rgba(153, 153, 153, 0.7);
-                cursor: pointer;
-              }
-            }
-          }
-        }
-      }
-    }
-
-    .selectSeatInfo {
-      display: grid;
-      grid-template-rows: 1fr 3fr 1fr 1fr;
-      background-color: $Gray33;
-
-      .text {
-        @include flexCenter;
-        font-size: 2rem;
-        color: white;
-      }
-
-      .selectSeatList {
-        margin: 10px;
-        padding: 20px;
-        border: 2px solid white;
-        font-size: 1.1rem;
-
-        .selectSeat {
-          color: white;
-        }
-      }
-
-      .totalPrice {
-        @include flexCenter;
-        font-size: 1.3rem;
-        color: white;
-      }
-
-      .goToPayment {
-        @include flexCenter;
-
-        .goToPaymentButton {
-          width: 80%;
-          height: 80%;
-          border: none;
-          border-radius: 30px;
-          font-weight: 400;
-          font-size: 1.5rem;
-          background-color: $Red;
-          color: white;
-          &:hover {
-            background-color: rgba(229, 8, 21, 0.7);
-            cursor: pointer;
-          }
-        }
-      }
-    }
 
     .orderFooter {
       grid-column: span 2;
